Revert "Moved the version class and regex into a new file"
This reverts commit f951746e55f93b662852ac35dd658527ca707a68.
